# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Login successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "forgot-password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Password reset request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Forgot password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset-requests": {
      "get": {
        "operationId": "list-password-reset-requests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Password reset requests retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PasswordResetRequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List password reset requests (Admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset-requests/{id}/approve": {
      "patch": {
        "operationId": "approve-password-reset-request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Request approved"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve password reset request (Admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset-requests/{id}/reject": {
      "patch": {
        "operationId": "reject-password-reset-request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Request rejected"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject password reset request (Admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "reset-password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "change-password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Change password request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid current password"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password-page": {
      "get": {
        "operationId": "reset-password-page",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Reset password token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reset password page (Development only)",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/verify": {
      "post": {
        "operationId": "verify-token",
        "parameters": [
          {
            "name": "authorization",
            "required": false,
            "in": "header",
            "description": "Optional Bearer token for non-cookie clients",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cookie",
            "required": false,
            "in": "header",
            "description": "Optional auth cookie for browser sessions",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          }
        },
        "summary": "Verify JWT token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/kpi/my-space/products/top-selling": {
      "post": {
        "operationId": "ResellerProductKpi_getTopSellingProducts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SellingProductDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get top selling products for a reseller",
        "tags": [
          "Reseller KPIs - Products"
        ]
      }
    },
    "/kpi/my-space/products/longest-time-in-inventory": {
      "post": {
        "operationId": "ResellerProductKpi_getProductsWithLongestTimeInInventory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductInInventoryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get products with longest time in inventory for a reseller",
        "tags": [
          "Reseller KPIs - Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luxis API",
    "description": "System API for Luxis",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PhoneNumber": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "The phone number (stored as digits only)",
            "example": "11987654321"
          }
        },
        "required": [
          "value"
        ]
      },
      "Email": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "The email address of the user"
          }
        },
        "required": [
          "value"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "test@test.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user. Must contain at least 10 characters, one uppercase letter, one lowercase letter, one number and one special character",
            "example": "Test@123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetRequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "$ref": "#/components/schemas/Email"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneNumber"
          },
          "token": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "COMPLETED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "approvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "rejectedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "userId",
          "username",
          "email",
          "phone",
          "token",
          "status",
          "createdAt"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Reset password token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "newPassword": {
            "type": "string",
            "description": "New password (must contain at least 10 characters, one uppercase, one lowercase, one number and one special character)",
            "example": "NewPassword123!",
            "minLength": 10
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "New password (must contain at least 10 characters, one uppercase, one lowercase, one number and one special character)",
            "example": "NewPassword123!",
            "minLength": 10
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "UserPayloadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the user",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "The role of the user",
            "example": "ADMIN",
            "enum": [
              "ADMIN",
              "RESELLER",
              "ASSISTANT",
              "UNASSIGNED"
            ],
            "default": "ADMIN"
          },
          "status": {
            "type": "string",
            "description": "The status of the user",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "PENDING"
            ],
            "default": "ACTIVE"
          },
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "name"
        ]
      },
      "VerifyDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "The verification token",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "user": {
            "description": "The user payload containing user details",
            "example": {
              "id": "123e4567-e89b-12d3-a456-426614174000",
              "email": "john.doe@example.com",
              "role": "ADMIN",
              "status": "ACTIVE",
              "name": "John Doe"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/UserPayloadDto"
              }
            ]
          }
        },
        "required": [
          "valid",
          "user"
        ]
      },
...
      "ParamsDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00Z"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00Z"
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "page": {
            "type": "number",
            "example": 1
          }
        }
      },
...
      "SellingProductDto": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "string",
            "description": "The ID of the product model",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "modelName": {
            "type": "string",
            "description": "The name of the product model",
            "example": "iPhone 13 Pro Max"
          },
          "quantity": {
            "type": "number",
            "description": "The quantity of the product",
            "example": 10
          },
          "salePrice": {
            "type": "string",
            "description": "The sale price of the product",
            "example": "150.00"
          },
          "totalValue": {
            "type": "string",
            "description": "The total value of the product",
            "example": "1500.00"
          }
        },
        "required": [
          "modelId",
          "modelName",
          "quantity",
          "salePrice",
          "totalValue"
        ]
      },
      "ProductInInventoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the product",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "serialNumber": {
            "type": "string",
            "description": "The serial number of the product",
            "example": "SN20240101001"
          },
          "modelId": {
            "type": "string",
            "description": "The ID of the product model",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "modelName": {
            "type": "string",
            "description": "The name of the product model",
            "example": "iPhone 13 Pro Max"
          },
          "salePrice": {
            "type": "string",
            "description": "The sale price of the product",
            "example": "150.00"
          },
          "dateAcquired": {
            "type": "string",
            "description": "The date when the product was acquired (batch arrival date)",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "modelId",
          "modelName",
          "salePrice",
          "dateAcquired"
        ]
      }
    }
  }
}